FileDownlink: Add file downlink queue and SendFile port
SendFile now adds files to a file downlink queue.
FileDownlink automatically pulls and downlinks files from the queue,
allowing users to queue up a list of files to downlink.

A SendFile port was added, allowing components to add files to the
file downlink queue. When a file downlink initated by a port completes,
a notification is broadcast on the FileComplete port, allowing components
to react when files they requested were downlinked.

In order to prevent a queue of files from saturating the communication link,
an cooldown state was added, allowing users to specify a delay between when a
file downlink completes and the next one begins.