Remove references to install command from the tutorial (#1877) (#1928)
The tutorial refers to the "install" command in the 'Build Ref
Deployment' section.  However, that command was removed in b31ae5c, a
commit released as early as v2.0.0~193^2.

This commit removes references to the install command from the tutorial.

The mention that the build command also copies the binaries to the bin
dir is moved earlier in the text, simply because it flows better when
the rest of the note about the install command is removed.